+++ src/main/java/mil/nga/geopackage/core/contents/Contents.java
@@ -117,7 +117,7 @@ public class Contents {
 	 * timestamp value in ISO 8601 format as defined by the strftime function
 	 * %Y-%m-%dT%H:%M:%fZ format string applied to the current time
 	 */
-	@DatabaseField(columnName = COLUMN_LAST_CHANGE, persisterClass = DatePersister.class)
+	@DatabaseField(columnName = COLUMN_LAST_CHANGE, persisterClass = DatePersister.class, defaultValue = DatePersister.DATE_FORMAT)
 	private Date lastChange;
 
 	/**